% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemio_sdo.R
\name{chemio_sdo}
\alias{chemio_sdo}
\title{Aggiunge al flusso sdo una variabile chemio_dia e cjemio_int corrispondenti ai ricoveri con codici chemioterapici}
\usage{
chemio_sdo(flusso_sdo)
}
\arguments{
\item{flusso_sdo}{}
}
\value{
aggiunge al flusso sdo una variabile chemio_dia e chemio_int corrispondenti ai ricoveri con codici chemioterapici V5811, V5812, 9225
}
\description{
Aggiunge al flusso sdo una variabile chemio_dia e cjemio_int corrispondenti ai ricoveri con codici chemioterapici
}
\examples{

}
