% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klung.R
\name{klung}
\alias{klung}
\title{Genera un vettore logico per selezionare le diagnosi di K polmone}
\usage{
klung(diagnosi_sdo)
}
\arguments{
\item{diagnosi_sdo}{vettore con codici diagnosi principale o secondaria}
}
\value{
valore logico in corrispondeza alle diagnosi principali o secondaria 1 162.x o 231.2
}
\description{
Genera un vettore logico per selezionare le diagnosi di K polmone
}
\examples{

}
