% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klungf.R
\name{klungf}
\alias{klungf}
\title{Genera un vettore con codici ATC}
\usage{
klungf(ATC)
}
\arguments{
\item{ATC}{vettore con codici ATC}
}
\value{
valore logico in corrispondeza ai farmaci crizotinib, alectinib,
osimertinib, gefitinib, erlotinib, afatinib, bevacizumab, nivolumab, pembrolizumab,
etoposide, topotecan, irinotecan, atezolizumab, durvalumab, ceritinib
}
\description{
Genera un vettore con codici ATC
}
\examples{

}
