% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca_data.R
\name{lca_data}
\alias{lca_data}
\title{Genera un dataframe da utilizzare con la funzione POLCA}
\usage{
lca_data(flusso_asa, flusso_farmaci, flusso_sdo)
}
\arguments{
\item{flusso_sdo}{}
}
\value{
Un database da utilizzare con la funzione POLCA
}
\description{
Genera un dataframe da utilizzare con la funzione POLCA
}
\examples{

}
