% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radio_sdo.R
\name{radio_sdo}
\alias{radio_sdo}
\title{aggiunge al flusso sdo una variabile radio_dia e radio_int corrispondenti ai ricoveri con codici radiologici}
\usage{
radio_sdo(flusso_sdo)
}
\arguments{
\item{flusso_sdo}{}
}
\value{
aggiunge al flusso sdo una variabile radio_dia e radio_int corrispondenti ai ricoveri con codici radiologici V580, 9229, 9221,
9222, 9223, 9224, 9225, 9226, 9227, 9228, 9229, 9230, 9231, 9232, 9239
}
\description{
aggiunge al flusso sdo una variabile radio_dia e radio_int corrispondenti ai ricoveri con codici radiologici
}
\examples{

}
