% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_asa.R
\name{select_asa}
\alias{select_asa}
\title{Genera un dataframe con prestazioni selezionate}
\usage{
select_asa(flusso_specialistica)
}
\arguments{
\item{flusso_specialistica}{}
}
\value{
database con prestazioni selezionate sui codici 87.03.1, 87.41, 87.41.1, 88.01.1, 88.01.2,
88.01.5, 88.01.6, 88.91.1 , 88.91.2 ,  92.11.7 ,   92.18.6,  99.25 ,   99.25.4 , 99.25.5 , 92.23.2 ,
92.24.01 , 92.24.02 ,92.24.1,  92.24.2 , 92.24.3 , 92.24.B , 92.25.2, 92.25.1
}
\description{
Genera un dataframe con prestazioni selezionate
}
\examples{

}
