% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_sdo.R
\name{select_sdo}
\alias{select_sdo}
\title{Genera  Un database con la selezione delle sdo sulla base degli interventi principale e secondatio (codici 31-35)}
\usage{
select_sdo(flusso_sdo)
}
\arguments{
\item{flusso_sdo}{}
}
\value{
Un database con la selezione delle sdo sulla base degli interventi principale e secondatio (codici 31-35) e crea una variabile intervento con ccodice 33
}
\description{
Genera  Un database con la selezione delle sdo sulla base degli interventi principale e secondatio (codici 31-35)
}
\examples{

}
